backend/pyproject.toml
[tool.poetry]
name = "caby-backend"
version = "0.1.0"
description = ""
authors = ["Caby Team"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlmodel = "^0.0.14"
asyncpg = "^0.29.0"
python-socketio = "^5.11.1"
redis = "^5.0.1"
celery = "^5.3.6"
googlemaps = "^4.10.0"
pydantic-settings = "^2.1.0"
geoalchemy2 = "^0.14.3"
shapely = "^2.0.2"
httpx = "^0.26.0"
stripe = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"